+++ src/lwplsrdaavg.jl
@@ -1,76 +1,91 @@
-"""
-    lwplsrdaavg(X, y; nlvdis, metric, h, k, nlv, 
-        tol = 1e-4, scal::Bool = false, verbose = false)
-Averaging of kNN-LWPLSR-DA models with different numbers of 
-    latent variables (LVs).
-* `X` : X-data.
-* `y` : y-data (class membership).
-* `nlvdis` : Number of latent variables (LVs) to consider in the global PLS 
-    used for the dimension reduction before calculating the dissimilarities. 
-    If `nlvdis = 0`, there is no dimension reduction.
-* `metric` : Type of dissimilarity used to select the neighbors. 
-    Possible values are :eucl (default; Euclidean distance) 
-    and :mah (Mahalanobis distance).
-* `h` : A scalar defining the shape of the weight function. Lower is h, 
-    sharper is the function. See function `wdist`.
-* `k` : The number of nearest neighbors to select for each observation to predict.
-* `nlv` : A character string such as "5:20" defining the range of the numbers of LVs 
-    to consider ("5:20": the predictions of models with nb LVS = 5, 6, ..., 20 
-    are averaged). Syntax such as "10" is also allowed ("10": correponds to 
-    the single model with 10 LVs).
-* `tol` : For stabilization when very close neighbors.
-* `scal` : Boolean. If `true`, each column of `X` 
-    is scaled by its uncorrected standard deviation.
-    The scaling is implemented for the global (distances) and local (i.e. inside
-    each neighborhood) computations.
-* `verbose` : If true, fitting information are printed.
+struct LwplsrdaAvg
+    X::Matrix
+    y::AbstractMatrix
+    fm
+    metric::String
+    h::Real
+    k::Int
+    nlv::String
+    tol::AbstractFloat
+    scal::Bool
+    verbose::Bool
+    lev::Vector
+    ni::Int
+end
 
-This is the same methodology as for `lwplsravg` except that 
-PLSR is replaced by PLSR-DA, and the mean is replaced by votes.
+#"""
+#    lwplsrdaavg(X, y; nlvdis, metric, h, k, nlv, 
+#        tol = 1e-4, scal::Bool = false, verbose = false)
+#Averaging of kNN-LWPLSR-DA models with different numbers of 
+#    latent variables (LVs).
+#* `X` : X-data.
+#* `y` : y-data (class membership).
+#* `nlvdis` : Number of latent variables (LVs) to consider in the global PLS 
+#    used for the dimension reduction before calculating the dissimilarities. 
+#    If `nlvdis = 0`, there is no dimension reduction.
+#* `metric` : Type of dissimilarity used to select the neighbors. 
+#    Possible values are :eucl (default; Euclidean distance) 
+#    and :mah (Mahalanobis distance).
+#* `h` : A scalar defining the shape of the weight function. Lower is h, 
+#    sharper is the function. See function `wdist`.
+#* `k` : The number of nearest neighbors to select for each observation to predict.
+#* `nlv` : A character string such as "5:20" defining the range of the numbers of LVs 
+#    to consider ("5:20": the predictions of models with nb LVS = 5, 6, ..., 20 
+#    are averaged). Syntax such as "10" is also allowed ("10": correponds to 
+#    the single model with 10 LVs).
+#* `tol` : For stabilization when very close neighbors.
+#* `scal` : Boolean. If `true`, each column of `X` 
+#    is scaled by its uncorrected standard deviation.
+#    The scaling is implemented for the global (distances) and local (i.e. inside
+#    each neighborhood) computations.
+#* `verbose` : If true, fitting information are printed.
 
-For instance, if argument `nlv` is set to `nlv = "5:10"`, the prediction for 
-a new observation is the most occurent class within the predictions 
-returned by the models with 5 LVS, 6 LVs, ... 10 LVs, respectively.
+#This is the same methodology as for `lwplsravg` except that 
+#PLSR is replaced by PLSR-DA, and the mean is replaced by votes.
 
-## Examples
-```julia
-using JLD2
-using JchemoData
-path_jdat = dirname(dirname(pathof(JchemoData)))
-db = joinpath(path_jdat, "data/forages2.jld2") 
-@load db dat
-pnames(dat)
+#For instance, if argument `nlv` is set to `nlv = "5:10"`, the prediction for 
+#a new observation is the most occurent class within the predictions 
+#returned by the models with 5 LVS, 6 LVs, ... 10 LVs, respectively.
 
-X = dat.X 
-Y = dat.Y 
-s = Bool.(Y.test)
-Xtrain = rmrow(X, s)
-ytrain = rmrow(Y.typ, s)
-Xtest = X[s, :]
-ytest = Y.typ[s]
+### Examples
+#```julia
+#using JLD2
+#using JchemoData
+#path_jdat = dirname(dirname(pathof(JchemoData)))
+#db = joinpath(path_jdat, "data/forages2.jld2") 
+#@load db dat
+#pnames(dat)
 
-tab(ytrain)
-tab(ytest)
+#X = dat.X 
+#Y = dat.Y 
+#s = Bool.(Y.test)
+#Xtrain = rmrow(X, s)
+#ytrain = rmrow(Y.typ, s)
+#Xtest = X[s, :]
+#ytest = Y.typ[s]
 
-nlvdis = 25 ; metric = :mah
-h = 2 ; k = 100
-nlv = "0:20"
-fm = lwplsrdaavg(Xtrain, ytrain;
-    nlvdis = nlvdis, metric = metric,
-    h = h, k = k, nlv = nlv) ;
-pnames(fm)
+#tab(ytrain)
+#tab(ytest)
 
-res = Jchemo.predict(fm, Xtest) ;
-pnames(res)
-res.pred
-err(res.pred, ytest)
-confusion(res.pred, ytest).cnt
+#nlvdis = 25 ; metric = :mah
+#h = 2 ; k = 100
+#nlv = "0:20"
+#fm = lwplsrdaavg(Xtrain, ytrain;
+#    nlvdis = nlvdis, metric = metric,
+#    h = h, k = k, nlv = nlv) ;
+#pnames(fm)
 
-res.listnn
-res.listd
-res.listw
-```
-""" 
+#res = Jchemo.predict(fm, Xtest) ;
+#pnames(res)
+#res.pred
+#err(res.pred, ytest)
+#confusion(res.pred, ytest).cnt
+
+#res.listnn
+#res.listd
+#res.listw
+#```
+#""" 
 function lwplsrdaavg(X, y; nlvdis, metric, h, k, nlv, 
     tol = 1e-4, scal::Bool = false, verbose = false)
     X = ensure_mat(X)
@@ -86,12 +101,12 @@ function lwplsrdaavg(X, y; nlvdis, metric, h, k, nlv,
         scal, verbose, ztab.keys, ztab.vals)
 end
 
-"""
-    predict(object::LwplsrdaAvg, X)
-Compute the y-predictions from the fitted model.
-* `object` : The fitted model.
-* `X` : X-data for which predictions are computed.
-""" 
+#"""
+#    predict(object::LwplsrdaAvg, X)
+#Compute the y-predictions from the fitted model.
+#* `object` : The fitted model.
+#* `X` : X-data for which predictions are computed.
+#""" 
 function predict(object::LwplsrdaAvg, X) 
     X = ensure_mat(X)
     m = nro(X)