+++ src/rfr_dt.jl
@@ -1,29 +1,41 @@
 """ 
-    treer_dt(X, y; n_subfeatures = 0,
+    rfr_dt(X, y; n_trees = 10,
+        partial_sampling = .7,  
+        n_subfeatures = -1,
         max_depth = -1, min_samples_leaf = 5, 
         min_samples_split = 2, scal::Bool = false, 
-        kwargs...)
-Regression tree (CART) with DecisionTree.jl.
+        mth = true, kwargs...)
+Random forest regression with DecisionTree.jl.
 * `X` : X-data (n obs., p variables).
 * `y` : Univariate y-data (n obs.).
-* `n_subfeatures` : Nb. variables to select at random at each split (default: 0 ==> keep all).
-* `max_depth` : Maximum depth of the decision tree (default: -1 ==> no maximum).
+* `n_trees` : Nb. trees built for the forest. 
+* `partial_sampling` : Proportion of sampled observations for each tree.
+* `n_subfeatures` : Nb. variables to select at random at each split (default: -1 ==> sqrt(#variables)).
+* `max_depth` : Maximum depth of the decision trees (default: -1 ==> no maximum).
 * `min_sample_leaf` : Minimum number of samples each leaf needs to have.
 * `min_sample_split` : Minimum number of observations in needed for a split.
 * `scal` : Boolean. If `true`, each column of `X` 
     is scaled by its uncorrected standard deviation.
-* `kwargs` : Optional named arguments to pass in function `build_tree` 
+* `mth` : Boolean indicating if a multi-threading is done when new data are 
+    predicted with function `predict`.
+* `kwargs` : Optional named arguments to pass in function `build_forest` 
     of `DecisionTree.jl`.
 
-The function fits a single regression tree (CART) using package 
+The function fits a random forest regression model using package 
 `DecisionTree.jl'.
 
 ## References
-Breiman, L., Friedman, J. H., Olshen, R. A., and Stone, C. J. Classification
-And Regression Trees. Chapman & Hall, 1984.
+Breiman, L., 1996. Bagging predictors. Mach Learn 24, 123–140. 
+https://doi.org/10.1007/BF00058655
+
+Breiman, L., 2001. Random Forests. Machine Learning 45, 5–32. 
+https://doi.org/10.1023/A:1010933404324
 
 DecisionTree.jl
 https://github.com/JuliaAI/DecisionTree.jl
+
+Genuer, R., 2010. Forêts aléatoires : aspects théoriques, 
+sélection de variables et applications. PhD Thesis. Université Paris Sud - Paris XI.
 
 Gey, S., 2002. Bornes de risque, détection de ruptures, boosting : 
 trois thèmes statistiques autour de CART en régression (These de doctorat). 
@@ -57,10 +69,8 @@ Xptrain = savgol(snv(Xtrain); f, pol, d)
 Xptest = savgol(snv(Xtest); f, pol, d) 
 
 n_subfeatures = p / 3 
-max_depth = 20
-fm = treer_dt(Xptrain, ytrain; 
-    n_subfeatures = n_subfeatures, 
-    max_depth = max_depth) ;
+fm = rfr_dt(Xptrain, ytrain; n_trees = 100,
+    n_subfeatures = n_subfeatures) ;
 pnames(fm)
 
 res = Jchemo.predict(fm, Xptest)
@@ -70,7 +80,7 @@ plotxy(res.pred, ytest; color = (:red, .5),
     bisect = true, xlabel = "Prediction", ylabel = "Observed").f  
 ```
 """ 
-function treer_dt(X, y; par) 
+function rfr_dt(X, y; par)
     X = ensure_mat(X)
     y = vec(y)
     p = nco(X)
@@ -81,41 +91,19 @@ function treer_dt(X, y; par)
     end
     n_subfeatures = Int(round(par.n_subfeatures))
     min_purity_increase = 0
-    fm = build_tree(y, X,
-        n_subfeatures,
-        par.max_depth,
+    fm = build_forest(y, X, 
+        n_subfeatures, 
+        par.n_trees, 
+        par.partial_sampling,
+        par.max_depth, 
         par.min_samples_leaf,
         par.min_samples_split,
         min_purity_increase;
         #kwargs...
         #rng = Random.GLOBAL_RNG
         #rng = 3
-        )
+        ) 
     featur = collect(1:p)
-    mth = true
-    TreerDt(fm, xscales, featur, mth)
+    TreerDt(fm, xscales, featur, par.mth)
 end
-
-"""
-    predict(object::TreerDt, X)
-Compute Y-predictions from a fitted model.
-* `object` : The fitted model.
-* `X` : X-data for which predictions are computed.
-""" 
-function predict(object::TreerDt, X)
-    X = ensure_mat(X)
-    m = nro(X)
-    if pnames(object.fm)[1] == :node
-        pred = apply_tree(object.fm, scale(X, object.xscales))
-    else
-        pred = apply_forest(object.fm, scale(X, object.xscales); 
-            use_multithreading = object.mth)
-    end
-    pred = reshape(pred, m, 1)
-    (pred = pred,)
-end
-
-
-
-